+++ app/get_cart_items.go
@@ -8,22 +8,22 @@ import (
 	log "github.com/sirupsen/logrus"
 )
 
-type itemsListResponseBody struct {
-	Object string                  `json:"object"`
-	Items  []*cartItemResponseBody `json:"items"`
-}
-
 // GetCartItemsHandler returns a list of all cart items
 func (a *App) GetCartItemsHandler() http.HandlerFunc {
+	type itemsListResponseBody struct {
+		Object string                  `json:"object"`
+		Items  []*cartItemResponseBody `json:"items"`
+	}
+
 	return func(w http.ResponseWriter, r *http.Request) {
 		ctx := r.Context()
 		contextLogger := log.WithContext(ctx)
 		contextLogger.Info("App: GetCartItemsHandler started")
 
-		uuid := chi.URLParam(r, "uuid")
-		cartItems, err := a.Service.GetCartItems(ctx, uuid)
+		id := chi.URLParam(r, "id")
+		cartItems, err := a.Service.GetCartItems(ctx, id)
 		if err != nil {
-			contextLogger.Errorf("service GetCartItems(%s) error: %v", uuid, err)
+			contextLogger.Errorf("service GetCartItems(%s) error: %v", id, err)
 			w.WriteHeader(http.StatusInternalServerError) // 500 Internal Server Error
 			return
 		}