+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # CHANGELOG
+## v0.54.0 (Tue 30 Jul 2019)
++ Cart routes `uuid` becomes `id`.
++ Cart operations return the `name` of the product alongside the `sku` and `qty` etc.
++ Schema test data uses `DESK-SKU=2987083`, `DRILL-SKU=395833`, `PHONE-SKU=241583`, `TV-SKU=2066250` and `WATER-SKU=20417`.
+
 ## v0.53.0 (Mon 29 Jul 2019)
 + Stripe webhook implemented
 + Stripe checkout implemented